Setup ElixirScript.
In order to not write JavaScript, ElixirScript is there.
Every module which should be translated to JavaScript needs to be
listed in the @js_modules module attribute of the
EphemeralWeb.Helper.JS module. Its 'call' method will then be
executed in the browser on pageload. WebAPIs are available once
mapped via ElixirScript.FFI, those mappings live in
lib/ephemeral_web/web_api.